+++ packages/core/tests/CLI.test.ts
@@ -1,4 +1,4 @@
-import { mockTool } from '@boost/test-utils';
+import { mockTool } from '../src/testUtils';
 import CLI from '../src/CLI';
 
 describe('CLI', () => {