+++ packages/core/tests/executors/Sync.test.ts
@@ -1,4 +1,4 @@
-import { mockTool } from '@boost/test-utils';
+import { mockTool } from '../../src/testUtils';
 import SyncExecutor from '../../src/executors/Sync';
 import Task from '../../src/Task';
 import Context from '../../src/Context';