+++ packages/core/tests/reporters/CIReporter.test.ts
@@ -1,5 +1,5 @@
 import chalk from 'chalk';
-import { mockTool, mockConsole } from '@boost/test-utils';
+import { mockTool, mockConsole } from '../../src/testUtils';
 import CIReporter from '../../src/reporters/CIReporter';
 
 describe('CIReporter', () => {