+++ packages/core/tests/reporters/ErrorReporter.test.ts
@@ -1,4 +1,4 @@
-import { mockTool, mockConsole } from '@boost/test-utils';
+import { mockTool, mockConsole } from '../../src/testUtils';
 import ErrorReporter from '../../src/reporters/ErrorReporter';
 
 describe('ErrorReporter', () => {